Pulled in r695 and and r699 from upstream LOVD3, to allow more transcripts. - LOVD now allows more transcripts to be stored; up to 16 million.

+++ src/install/inc-sql-tables.php
@@ -4,10 +4,10 @@
  * LEIDEN OPEN VARIATION DATABASE (LOVD)
  *
  * Created     : 2009-10-22
- * Modified    : 2014-06-16
- * For LOVD    : 3.0-11
+ * Modified    : 2015-12-03
+ * For LOVD    : 3.0-15
  *
- * Copyright   : 2004-2014 Leiden University Medical Center; http://www.LUMC.nl/
+ * Copyright   : 2004-2015 Leiden University Medical Center; http://www.LUMC.nl/
  * Programmers : Ing. Ivo F.A.C. Fokkema <I.F.A.C.Fokkema@LUMC.nl>
  *               Ing. Ivar C. Lugtenburg <I.C.Lugtenburg@LUMC.nl>
  *
@@ -32,6 +32,19 @@
 // STILL TODO:
 // variant <-> pathogenicity <-> disease? Link pathogenicity specifically to one of the phenotypes or diseases?
 // Functional assays / computer predictions, hoe toevoegen??? Aan variant én aan individual???
+
+// IDs:
+// WARNING: If editing any of these, also edit $_SETT['objectid_length']!
+// userid SMALLINT(5) UNSIGNED (65K)
+// geneid VARCHAR(25) (25 characters)
+// transcriptid MEDIUMINT(8) UNSIGNED (16M)
+// diseaseid SMALLINT(5) UNSIGNED (65K)
+// individualid MEDIUMINT(8) UNSIGNED (16M)
+// variantid INT(10) UNSIGNED (4294M)
+// phenotypeid INT(10) UNSIGNED (4294M)
+// screeningid INT(10) UNSIGNED (4294M)
+// colid VARCHAR(100) (100 characters)
+// linkid TINYINT(3) UNSIGNED (255)
 
 // DMD_SPECIFIC
 if (!defined('ROOT_PATH')) {
@@ -161,7 +174,7 @@ $aTableSQL =
 
          , 'TABLE_TRANSCRIPTS' =>
    'CREATE TABLE ' . TABLE_TRANSCRIPTS . ' (
-    id SMALLINT(5) UNSIGNED ZEROFILL NOT NULL AUTO_INCREMENT,
+    id MEDIUMINT(8) UNSIGNED ZEROFILL NOT NULL AUTO_INCREMENT,
     geneid VARCHAR(25) NOT NULL,
     name VARCHAR(255) NOT NULL,
     id_mutalyzer TINYINT(3) UNSIGNED ZEROFILL,
@@ -382,7 +395,7 @@ $aTableSQL =
          , 'TABLE_VARIANTS_ON_TRANSCRIPTS' =>
    'CREATE TABLE ' . TABLE_VARIANTS_ON_TRANSCRIPTS . ' (
     id INT(10) UNSIGNED ZEROFILL NOT NULL,
-    transcriptid SMALLINT(5) UNSIGNED ZEROFILL NOT NULL,
+    transcriptid MEDIUMINT(8) UNSIGNED ZEROFILL NOT NULL,
     effectid TINYINT(2) UNSIGNED ZEROFILL,
     position_c_start MEDIUMINT,
     position_c_start_intron INT,
